added serialization of pages
