+++ Resources/config/serializer/Document.ArticleDocument.xml
@@ -26,7 +26,7 @@
         <property name="structure" type="Sulu\Component\Content\Document\Structure\Structure" groups="preview"/>
         <property name="structureType" type="string" groups="preview"/>
 
-        <property name="title" type="string" groups="defaultArticle,smallArticle,preview"/>
+        <property name="title" type="string" groups="website,defaultArticle,smallArticle,preview"/>
         <property name="creator" type="integer" groups="website,defaultArticle,preview"/>
         <property name="changer" type="integer" groups="website,defaultArticle,preview"/>
         <property name="created" type="DateTime" groups="website,defaultArticle,preview"/>
