+++ Resources/config/serializer/Document.ArticlePageDocument.xml
@@ -27,7 +27,7 @@
         <property name="structure" type="Sulu\Component\Content\Document\Structure\Structure" groups="preview"/>
         <property name="structureType" type="string" groups="preview"/>
 
-        <property name="title" serialized-name="articleTitle" type="string" groups="preview"/>
+        <property name="title" serialized-name="articleTitle" type="string" groups="website,preview"/>
         <property name="pageNumber" type="integer" groups="website,defaultArticlePage,smallArticlePage,preview"/>
         <property name="creator" type="integer" groups="website,defaultArticlePage,smallArticlePage,preview"/>
         <property name="changer" type="integer" groups="website,defaultArticlePage,smallArticlePage,preview"/>
