+++ Resources/config/services.xml
@@ -145,8 +145,16 @@
             <!-- This needs to happen after the content repository has been initialized !-->
             <tag name="sulu_document_manager.initializer" priority="-127"/>
         </service>
+
         <service id="sulu_article.serializer.article"
                  class="Sulu\Bundle\ArticleBundle\Document\Serializer\ArticleSubscriber">
+            <argument type="service" id="sulu.content.structure_manager"/>
+
+            <tag name="jms_serializer.event_subscriber" />
+        </service>
+
+        <service id="sulu_article.serializer.website_article"
+                 class="Sulu\Bundle\ArticleBundle\Document\Serializer\ArticleWebsiteSubscriber">
             <argument type="service" id="sulu.content.structure_manager"/>
             <argument type="service" id="sulu.content.type_manager"/>
             <argument type="service" id="sulu_article.content.data_provider.proxy_factory"/>
